refactor: move config & init functions to a separate folder
