+++ packages/cli/src/commands/frameworks/index.ts
@@ -1,0 +1,9 @@
+import loadConfigVue from './vue/config'
+import initVue from './vue/init'
+
+export default {
+  vue: {
+    loadConfig: loadConfigVue,
+    init: initVue,
+  },
+}
